+++ test/unittests/t_cache.cc
@@ -242,7 +242,7 @@ class TestCacheManager : public CacheManager {
   }
   virtual int Dup(int fd) { return fd; }
   virtual int Readahead(int fd) { return 0; }
-  virtual uint16_t SizeOfTxn() { return sizeof(int); }
+  virtual uint32_t SizeOfTxn() { return sizeof(int); }
   virtual int StartTxn(const shash::Any &id, uint64_t size, void *txn) {
     int fd = open("/dev/null", O_RDONLY);
     assert(fd >= 0);
